% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CanadianBigData.R
\docType{package}
\name{CanadianBigData}
\alias{CanadianBigData}
\title{CanadianOralOpioids: TODO description, e.g., Obtain the latest information on Morphine administration provided by HealthCanada
#TODO find better package name with opiod in name}
