% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CanadianOralOpioids.R
\docType{package}
\name{CanadianOralOpioids}
\alias{CanadianOralOpioids}
\title{CanadianOralOpioids: Obtain the latest information on Morphine equivalent doses provided by HealthCanada}
\description{
This package provides details on Oral Opioids approved for sale by Health Canada.
Please note that the output generated by the package should not be substituted for clinical
advise and any medication should be only consumed at the advise of a licensed healthcare provider.
}
