% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CanadianBigData.R
\name{load_Big_Data_form}
\alias{load_Big_Data_form}
\title{Obtain the latest Opiod data from Health Canada}
\usage{
load_Big_Data_form(filelocation = "", verbose = TRUE)
}
\arguments{
\item{filelocation}{The directory on your system where you want the dataset to be downloaded.
If "", filelocation will be set to the download path within the CanadianBigData
package installation directory.}

\item{verbose}{Logical. Indicates whether messages will be printed in the console. Default: \code{TRUE}.}
}
\value{
The function returns a Boolean for the update status. Attributes
include a status message (msg). The dataset showing all details of all oral
opioids available for sale as authorized by HealthCanada (Big_data_form),
and its save path (path).
}
\description{
\code{load_Big_Data_form} compares the date of the local Big_Data_form and compares
it with the latest date of data provided Health Canada. In case the local file is outdated,
an updated file will be generated.
}
\examples{
load_Big_Data_form()
}
