% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CanadianBigData.R
\name{MED_50}
\alias{MED_50}
\title{Get the Morphine Equivalent Dose (MED) from Health Canada by using the DIN}
\usage{
MED_50(Drug_ID, HealthCanada_Opioid_Table)
}
\arguments{
\item{Drug_ID}{A numeric value for the DIN. Exclude all zeros in front}

\item{HealthCanada_Opioid_Table}{Health Canada opiod dataset which can be loaded by using
the \code{load_HealthCanada_Opioid_Table()} function.}
}
\value{
MED: Morphine Equivalent Dose
}
\description{
\code{MED} retrieves the Morphine Equivalent Dose from the HealthCanada_Opioid_Table.
}
\examples{
MED_50(108316, HealthCanada_Opioid_Table)
}
