% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CanadianBigData.R
\name{Opioid}
\alias{Opioid}
\title{Get the Morphine Equivalent Dose (MED) from Health Canada by using the DIN}
\usage{
Opioid(Drug_ID, Big_Data_form)
}
\arguments{
\item{Drug_ID}{A numeric value for the DIN. Exclude all zeros in front}

\item{Big_Data_form}{Health Canada opiod dataset which can be loaded by using
the \code{load_Big_Data_form()} function.}
}
\value{
MED: Morphine Equivalent Dose
}
\description{
\code{MED} retrieves the Morphine Equivalent Dose from the Big_Data_form.
}
\examples{
Opioid(108316, Big_data_form)
}
