% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OralOpioids.R
\docType{package}
\name{OralOpioids-package}
\alias{OralOpioids}
\alias{OralOpioids-package}
\title{OralOpioids: Obtain the latest information on Morphine equivalent doses provided by HealthCanada and the FDA}
\description{
This package provides details on Oral Opioids approved for sale by Health Canada and the FDA.
Please note that the output generated by the package should not be substituted for clinical
advise and any medication should be only consumed at the advise of a licensed healthcare provider.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ankonahouston/OralOpioids}
}

}
\author{
\strong{Maintainer}: Ankona Banerjee \email{ankonanagpur@gmail.com} (\href{https://orcid.org/0000-0003-4433-5446}{ORCID})

Authors:
\itemize{
  \item Erik Stricker \email{erik.stricker@gmx.com} (\href{https://orcid.org/0000-0003-2112-2202}{ORCID})
}

}
\keyword{internal}
