% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CanadianBigData.R
\name{load_HealthCanada_Opioid_Table}
\alias{load_HealthCanada_Opioid_Table}
\title{Obtain the latest Opiod data from Health Canada}
\usage{
load_HealthCanada_Opioid_Table(
  filelocation = "",
  no_download = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{filelocation}{String. The directory on your system where you want the dataset to be downloaded.
If "", filelocation will be set to the download path within the CanadianBigData
package installation directory.}

\item{no_download}{Logical. If set to TRUE, no downloads will be executed and no user input is required. Default: \code{FALSE}.}

\item{verbose}{Logical. Indicates whether messages will be printed in the console. Default: \code{TRUE}.}
}
\value{
The function returns a Boolean for the update status. Attributes
include a status message (msg). The dataset showing all details of all oral
opioids available for sale as authorized by HealthCanada (HealthCanada_Opioid_Table),
and its save path (path).
}
\description{
\code{load_HealthCanada_Opioid_Table} compares the date of the local HealthCanada_Opioid_Table and compares
it with the latest date of data provided Health Canada. In case the local file is outdated,
an updated file will be generated.
}
\examples{
load_HealthCanada_Opioid_Table(no_download = TRUE)
}
